+++ plugin_ide.core/src-lang/melnorme/lang/ide/launching/LaunchConstants_Default.java
@@ -6,6 +6,7 @@ public interface LaunchConstants_Default {
 	public static final String ATTR_BUILD_TARGET = "buildTarget";
 	public static final String ATTR_PROGRAM_PATH_USE_DEFAULT = "programPath_UseDefault";
 	public static final String ATTR_PROGRAM_PATH = "programPath";
+	
 	public static final String ATTR_PROGRAM_ARGUMENTS = "programArguments";
 	public static final String ATTR_WORKING_DIRECTORY = "workingDir";